requirements.txt
# Chatterbox TTS Server - Requirements
# Optimized for L40S GPU with CUDA support

# Core inference
onnxruntime-gpu>=1.16.0
numpy>=1.24.0
transformers>=4.35.0
huggingface_hub>=0.19.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# API server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Benchmarking
httpx>=0.25.0
tqdm>=4.66.0

# Optional: Watermarking
# perth

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
